refactor user profile view
